add config flattening functionality

+++ src/main/scala/com/typesafe/config/impl/Internals.scala
@@ -12,7 +12,7 @@ object Internals {
 
   def parseDuration(x: String, o: ConfigOrigin, p: String): Long = SimpleConfig.parseDuration(x, o, p)
 
-  def parseBytes(x: String, o: ConfigOrigin, p: String): Long = SimpleConfig.parseBytes(x, o, p)
+  def parseBytes(x: String, o: ConfigOrigin, p: String): Long = SimpleConfig.parseBytes(x, o, p).longValue()
 
   def parsePeriod(x: String, o: ConfigOrigin, p: String): Period = SimpleConfig.parsePeriod(x, o, p)
 }
